Add files via upload
The fiscal multiplier is a central instrument of policy makers. Understanding its magnitude and effect in the economy is imperative to the correct implementation of fiscal policy. In this article we depict the Antonio Acconciaa, Giancarlo Corsetti and Saverio Simonelli 2014 article, Mafia and Public Spending: Evidence on the Fiscal Multiplier from a Quasi-Experiment, evaluating their results in order to understand how in times of high criminality a budget cut could impact economies output, and obtaining a fiscal multiplier estimate for Italy. 

Please find the prensentation: https://rpubs.com/nakerman95/778254